Fix merge mistake
