--- e2e/clients/react/index.js
@@ -1,8 +1,0 @@
-/* eslint-disable react/react-in-jsx-scope */
-import { createRoot } from 'react-dom/client'
-import App from './App'
-
-const container = document.getElementById('app')
-const root = createRoot(container)
-
-root.render(<App />)
