+++ e2e/clients/react/index.jsx
@@ -1,5 +1,8 @@
 /* eslint-disable react/react-in-jsx-scope */
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 
-ReactDOM.render(<App />, document.getElementById('app'))
+const container = document.getElementById('app')
+const root = createRoot(container)
+
+root.render(<App />)
